Fix some more
