+++ examples/wallet/src/types.ts
@@ -1,0 +1,10 @@
+import { IdentityProvider } from "@concordium/web-sdk";
+
+export interface IdentityProviderMetaData {
+    issuanceStart: string;
+    recoveryStart: string;
+    icon: string;
+    support: string;
+}
+
+export type IdentityProviderWithMetadata = IdentityProvider & { metadata: IdentityProviderMetaData };
